requirements.txt
flask
flask_login
werkzeug
flask_sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
psycopg2
